# Speed up `scipy.signal.stft` by using `lax.dynamic_slice_in_dim` for windowing