@dsuedholt 2025-05-12
Thanks for the review! I made the requested changes and squashed them into one commit.